Materials: rename materials.py into rendermaterials.py
