+++ Render.py
@@ -64,7 +64,7 @@ except ImportError:
 from renderutils import translate, debug, getproxyattr
 import camera
 import lights
-import materials
+import rendermaterials
 import renderables
 
 
@@ -939,7 +939,7 @@ class RendererHandler:
         write_object = functools.partial(RendererHandler._call_renderer,
                                          self,
                                          "write_object")
-        get_mat = functools.partial(materials.get_rendering_material,
+        get_mat = functools.partial(rendermaterials.get_rendering_material,
                                     renderer=self.renderer_name,
                                     default_color=view.Proxy.get_shape_color())
         res = [write_object(r.name, r.mesh, get_mat(r.material))
